renamed unitSpec into unit
